Next config template (#125)
