+++ .changeset/spotty-carpets-enjoy.md
@@ -1,0 +1,5 @@
+---
+"create-eth": patch
+---
+
+added nextjs config template
